requirements.txt
anyio==3.6.2
bcrypt==4.0.1
certifi==2023.5.7
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
email-validator==2.0.0.post2
fastapi==0.95.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.17.0
httptools==0.5.0
httpx==0.24.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
orjson==3.8.12
pandas==1.5.3
paramiko==2.6.0
pexpect==4.6.0
Pillow==9.5.0
platformdirs==3.5.0
protobuf==3.6.1
ptyprocess==0.7.0
pycparser==2.21
pydantic==1.10.7
PyNaCl==1.5.0
pyodbc==4.0.39
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.46
starlette==0.26.1
typing_extensions==4.5.0
tzdata==2023.3
ujson==5.7.0
urllib3==2.0.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
